Moved Licenses view model
